make run_occ21_test not delete .tce when passed -v

+++ tools/occ21/tests/run_occ21_test.py
@@ -71,11 +71,12 @@ def compile(filename):
                     else:
                         print '         no %s found on line %d' % (type.lower(), line)
 
-        try:
-            filename_tce = os.path.splitext(filename)[0] + '.tce'
-            os.remove(filename_tce)
-        except OSError, e:
-            pass
+        if not verbose:
+            try:
+                filename_tce = os.path.splitext(filename)[0] + '.tce'
+                os.remove(filename_tce)
+            except OSError, e:
+                pass
     else:
         test_result = 1
     # print useful info
